requirements.txt
aiohttp==3.9.5 ; python_full_version >= "3.8.1" and python_version < "4.0"
aiosignal==1.3.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
annotated-types==0.7.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
anyio==4.4.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
async-timeout==4.0.3 ; python_full_version >= "3.8.1" and python_version < "3.11"
attrs==23.2.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
broadcast-service==1.3.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
certifi==2024.6.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
click==8.1.7 ; python_full_version >= "3.8.1" and python_version < "4.0"
colorama==0.4.6 ; python_full_version >= "3.8.1" and python_version < "4.0" and platform_system == "Windows"
cushy-storage==1.3.8 ; python_full_version >= "3.8.1" and python_version < "4.0"
distro==1.9.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
exceptiongroup==1.2.1 ; python_full_version >= "3.8.1" and python_version < "3.11"
filelock==3.15.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
frozenlist==1.4.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
fsspec==2024.6.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
h11==0.14.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
httpcore==1.0.5 ; python_full_version >= "3.8.1" and python_version < "4.0"
httpx==0.27.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
huggingface-hub==0.23.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
idna==3.7 ; python_full_version >= "3.8.1" and python_version < "4.0"
ijson==3.3.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
importlib-metadata==7.2.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
jinja2==3.1.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
litellm==1.40.25 ; python_full_version >= "3.8.1" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
markupsafe==2.1.5 ; python_full_version >= "3.8.1" and python_version < "4.0"
mdurl==0.1.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
multidict==6.0.5 ; python_full_version >= "3.8.1" and python_version < "4.0"
numexpr==2.8.6 ; python_full_version >= "3.8.1" and python_version < "4.0"
numpy==1.24.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
openai==1.35.3 ; python_full_version >= "3.8.1" and python_version < "4.0"
packaging==24.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
pne==1.0.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
prompt-toolkit==3.0.36 ; python_full_version >= "3.8.1" and python_version < "4.0"
promptulate==1.16.7 ; python_full_version >= "3.8.1" and python_version < "4.0"
pydantic-core==2.18.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
pydantic==2.7.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
pygments==2.18.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
python-dotenv==1.0.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
pyyaml==6.0.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
questionary==2.0.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
regex==2024.5.15 ; python_full_version >= "3.8.1" and python_version < "4.0"
requests==2.32.3 ; python_full_version >= "3.8.1" and python_version < "4.0"
rich==13.7.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
shellingham==1.5.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
sniffio==1.3.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
tiktoken==0.7.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
tokenizers==0.19.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
tqdm==4.66.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
typer==0.12.3 ; python_full_version >= "3.8.1" and python_version < "4.0"
typing-extensions==4.12.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
urllib3==2.2.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
wcwidth==0.2.13 ; python_full_version >= "3.8.1" and python_version < "4.0"
yarl==1.9.4 ; python_full_version >= "3.8.1" and python_version < "4.0"
zipp==3.19.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
